resource: Move sources of the same topic into their own folders
Refactor to improve extendibility.

Move resource data and graph into "schema" directory

Move DFU traverser code into "traversers" directory, the folder
where different types of graph traverser can be introduced.

Move dfu_match_cb.[h|c]cpp and matcher.[h|c]pp into "policies/base"
directory.

Move scoring_api and other classes used for resource vertex
and edge evaluation into the "evaluators" folder.

Move resource generator code into "generators" directory and
shorten the file names.
