+++ resource/Makefile.am
@@ -24,8 +24,8 @@ EXTRA_DIST= \
 #
 utilities_grug2dot_SOURCES = \
     utilities/grug2dot.cpp \
-    resource_gen_spec.cpp \
-    resource_gen_spec.hpp
+    generators/spec.cpp \
+    generators/spec.hpp
 utilities_grug2dot_CXXFLAGS = \
     $(AM_CXXFLAGS)
 utilities_grug2dot_LDADD = \
@@ -43,38 +43,38 @@ utilities_resource_query_SOURCES = \
     policies/dfu_match_high_id_first.cpp \
     policies/dfu_match_low_id_first.cpp \
     policies/dfu_match_locality.cpp \
-    resource_data.cpp \
-    infra_data.cpp \
-    sched_data.cpp \
-    dfu_traverse.cpp \
-    dfu_traverse_impl.cpp \
-    dfu_match_cb.cpp \
-    matcher.cpp \
-    resource_gen.cpp \
-    resource_gen_spec.cpp \
-    color.cpp \
-    scoring_api.cpp \
-    edge_eval_api.cpp \
+    schema/resource_data.cpp \
+    schema/infra_data.cpp \
+    schema/sched_data.cpp \
+    schema/color.cpp \
+    traversers/dfu.cpp \
+    traversers/dfu_impl.cpp \
+    policies/base/dfu_match_cb.cpp \
+    policies/base/matcher.cpp \
+    generators/gen.cpp \
+    generators/spec.cpp \
+    evaluators/scoring_api.cpp \
+    evaluators/edge_eval_api.cpp \
     utilities/command.hpp \
     policies/dfu_match_high_id_first.hpp \
     policies/dfu_match_low_id_first.hpp \
     policies/dfu_match_locality.hpp \
-    resource_graph.hpp \
-    data_std.hpp \
-    infra_data.hpp \
-    sched_data.hpp \
-    resource_data.hpp \
-    resource_gen.hpp \
-    resource_gen_spec.hpp \
-    color.hpp \
-    dfu_traverse.hpp \
-    dfu_traverse_impl.hpp \
-    dfu_match_cb.hpp \
-    matcher.hpp \
-    system_defaults.hpp \
-    scoring_api.hpp \
-    edge_eval_api.hpp \
-    fold.hpp \
+    schema/resource_graph.hpp \
+    schema/data_std.hpp \
+    schema/infra_data.hpp \
+    schema/sched_data.hpp \
+    schema/resource_data.hpp \
+    schema/color.hpp \
+    traversers/dfu.hpp \
+    traversers/dfu_impl.hpp \
+    policies/base/dfu_match_cb.hpp \
+    policies/base/matcher.hpp \
+    generators/gen.hpp \
+    generators/spec.hpp \
+    evaluators/scoring_api.hpp \
+    evaluators/edge_eval_api.hpp \
+    evaluators/fold.hpp \
+    config/system_defaults.hpp \
     planner/planner.h
 utilities_resource_query_CXXFLAGS = \
     $(AM_CXXFLAGS) \
