+++ resource/schema/color.hpp
