+++ resource/schema/infra_data.hpp
@@ -25,7 +25,7 @@
 
 #include <map>
 #include <cstdint>
-#include "data_std.hpp"
+#include "schema/data_std.hpp"
 #include "planner/planner.h"
 
 namespace Flux {
