+++ resource/traversers/dfu_impl.cpp
@@ -22,7 +22,7 @@
  *  See also:  http://www.gnu.org/licenses/
 \*****************************************************************************/
 
-#include "dfu_traverse_impl.hpp"
+#include "traversers/dfu_impl.hpp"
 
 extern "C" {
 #if HAVE_CONFIG_H
